refactors code

+++ java/org/apache/juli/FileHandler.java
@@ -20,8 +20,10 @@ package org.apache.juli;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.FileNotFoundException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
@@ -52,7 +54,6 @@ import java.util.logging.LogRecord;
 import java.util.regex.Pattern;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
-import java.io.FileInputStream;
 import java.io.FileOutputStream;
 
 /**
@@ -102,6 +103,11 @@ public class FileHandler extends Handler {
 
     public static final int DEFAULT_MAX_DAYS = -1;
     public static final int DEFAULT_BUFFER_SIZE = -1;
+
+    /**
+     * The extension for the compressed rotated log file.
+     */
+    private static final String COMPRESSED_ROTATED_LOG_FILE_EXTENSION = ".gz";
 
 
     private static final ExecutorService DELETE_FILES_SERVICE =
@@ -347,10 +353,11 @@ public class FileHandler extends Handler {
             writer.write(getFormatter().getTail(this));
             writer.flush();
             writer.close();
-			if(compress && rotatable) {
-				compress();
-				deleteRotatedFile();
-			}
+
+            if(compress && rotatable) {
+                compressAndDelete();
+            }
+
 			writer = null;
             date = "";
         } catch (Exception e) {
@@ -359,45 +366,69 @@ public class FileHandler extends Handler {
             writerLock.writeLock().unlock();
         }
     }
-	
-	
-		
+
     /**
-     * Compress the file during the rotation.
+     * Compresses the rotated log file and deletes it to reduce storage space.
+     * <p>
+     * Check: https://bz.apache.org/bugzilla/show_bug.cgi?id=62611
      */
-	public synchronized void compress(){
-		    File dir = new File(directory);
-		    File currentLogFile = new File(dir.getAbsoluteFile(), prefix
-                    + (rotatable.booleanValue() ? date : "") + suffix);
-			byte[] buffer = new byte[1024];
-			try {
-				GZIPOutputStream out = new GZIPOutputStream(new FileOutputStream(currentLogFile+".gz"));
-				FileInputStream in = new FileInputStream(currentLogFile);
-                int lenght;
-                while((lenght=in.read(buffer)) != -1) {
-                    out.write(buffer, 0, lenght);
+    private synchronized void compressAndDelete() {
+        File currentDirectory = new File(this.directory);
+        File currentLogFile = new File(currentDirectory.getAbsoluteFile(), prefix
+                + (rotatable.booleanValue() ? date : "") + suffix);
+        if(currentLogFile.exists()) {
+            boolean compress = compressRotatedLogFile(currentLogFile);
+            if (compress) {
+                // TODO: original rotated file should probably be kept for a few days until deletion.
+                if (!currentLogFile.delete()) {
+                    reportError("Unable to delete current log file", null, ErrorManager.GENERIC_FAILURE);
                 }
-				in.close();
-				out.finish();
-				out.close();
-            } catch (IOException e) {
-                reportError("Unable to rotate log file", null,
-                        ErrorManager.GENERIC_FAILURE);
             }
         }
-    	
-	/**
-     * delete the file after is bieng compressed.
+    }
+
+    /**
+     *  Compresses the rotated log file.
+     *
+     *  @return true if the compression succeeded, otherwise, it return false.
      */
-	private void deleteRotatedFile() {
-		File dir = new File(directory);
-		File currentLogFile = new File(dir.getAbsoluteFile(), prefix
-                    + (rotatable.booleanValue() ? date : "") + suffix);
-        if (!currentLogFile.delete()) {
-            reportError("Unable to delete current log file", null,
-                        ErrorManager.GENERIC_FAILURE);
+    private boolean compressRotatedLogFile(File logFile){
+
+        // try to get handle of the currentLogFile.
+        FileInputStream currentLogFileInputStream = null;
+        try {
+            currentLogFileInputStream = new FileInputStream(logFile);
+        } catch (FileNotFoundException e) {
+            reportError("Unable to read from the current log file", e, ErrorManager.GENERIC_FAILURE);
+            return false;
         }
-        return;
+
+        // Create a handle to store the compressed rotated log file.
+        String compressedRotatedLogFilePath = logFile.getAbsolutePath() + COMPRESSED_ROTATED_LOG_FILE_EXTENSION;
+        FileOutputStream compressedRotatedLogFileOutputStream = null;
+        try {
+            compressedRotatedLogFileOutputStream = new FileOutputStream(compressedRotatedLogFilePath);
+        } catch (FileNotFoundException e) {
+            reportError("Unable to construct compressed log file", e, ErrorManager.GENERIC_FAILURE);
+            return false;
+        }
+
+        // Perform compression.
+        try {
+            GZIPOutputStream compressedRotatedLogFileGZIPOutputStream = new GZIPOutputStream(compressedRotatedLogFileOutputStream);
+            byte[] buffer = new byte[1024];
+            int length = 0;
+            while((length = currentLogFileInputStream.read(buffer)) != -1) {
+                compressedRotatedLogFileGZIPOutputStream.write(buffer, 0, length);
+            }
+            currentLogFileInputStream.close();
+            compressedRotatedLogFileGZIPOutputStream.finish();
+            compressedRotatedLogFileGZIPOutputStream.close();
+        } catch (IOException e) {
+            reportError("Unable to write to the compressed rotated log file", e, ErrorManager.GENERIC_FAILURE);
+            return false;
+        }
+        return true;
     }
 	
     /**
